Version 0.5
Now works for domains with an interface between two regions. 
Changed the output from ToLinearMatrixForm to not return the vectors from boundary conditions in the case of the homogeneous conditions, as this makes the CompoundMatrixMethod function simpler and only homogeneous equations can work in the CMM. Will output a warning for inhomogeneous conditions.